backend/requirement.txt
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
pytz==2024.2
questdb==2.0.3
questdb-connect==1.1.3
typing_extensions==4.12.2
